pyyaml/requirements.txt
Flask==3.0.0
PyYAML==6.0.1
Werkzeug==3.1.4

